tabs: re-write of basic tabs API
This commit is enormous.

This commit re-writes the basic tabs API to provide a single access point for all tab objects, the `Tab` class. This is built upon a metaclass and a bunch of flavour objects,

- StaticTab
- IntervalTab
- EventTab

based on the relationship between the tab to be created and GPS timing, read the code for details.