Use try-except when reading in an archive HDF5 file (#346)
* Use try-except when reading in an archive HDF5 file

Sometimes the HDF5 file is corrupted on the cluster and it causes the summary page job to fail. The only way to recover is to remove the corrupted file and then resubmit the failed job.

This change tries to open the file, but if there is an OSError raised, then the source file can be removed before continuing. Since no data is loaded, it is as if the file doesn't exist anyway. I think this should solve our recurring issues on the cluster regarding corrupted HDF5 files

Closes #344

Co-authored-by: Evan Goetz <evan.goetz@ligo.org>