Bugfix for when hours are requested in a state
Currently the "hours" option in a state is broken because the value for the day start is not being saved in the code. Basically the replace() method for a datetime object does not change the value in-place, it just returns the value. So essentially, all the broken code was doing was stating the day start but not actually using it. This fixes the problem