Added a new "metastate" section type where one can combine states from different detectors
This fixes a problem where the "Quiet" and "Noisy" states were undefined in time, so the whole duration was used. This meant there was no distinction between the states as plotted for the summary pages.

A metastate would be defined as
```
[metastate-Quiet]
key = Quiet
name = SEI Quiet
uses = H1-quiet,L1-quiet

[state-H1-Quiet]
name = SEI Quiet
key = H1-quiet
hours = 1-5,H1

[state-L1-Quiet]
name = SEI Quiet
key = L1-quiet
hours = 2-6,L1
```

It is not super robust in that it's really only supposed to be used for combining some interval for H1 and some interval for L1. I'm not sure it's really ready to be heavily used in many situations

Addresses the comment in #358